從選項式修正成組合式
