+++ src/views/Faq.vue
@@ -3,7 +3,7 @@ import { onMounted, ref, computed, watch } from 'vue'
 import axios from 'axios'
 import SidebarGrid from '@/components/SidebarGrid.vue'
 import router from '../router'
-import Notice from '../components/notification/notice.vue'
+import Notice from '../components/notification/Notice.vue'
 import NavLoginBtn from '../components/NavLoginBtn.vue'
 import MainFooter from '../components/MainFooter.vue'
